# openapi_dump.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HRMS API",
    "description": "HRMS Backend API \u2014 Enterprise Human Resource Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/attendance/check-in": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Employee Check-In",
        "description": "Clock in for today with server-side geofencing and IP validation.",
        "operationId": "check_in_api_v1_attendance_check_in_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/attendance/check-out": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Employee Check-Out",
        "description": "Clock out for today and compute total hours.",
        "operationId": "check_out_api_v1_attendance_check_out_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckOutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckOutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/attendance/me": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "Get current employee's monthly attendance",
        "description": "Retrieve monthly attendance history for currently logged-in employee.",
        "operationId": "get_my_attendance_api_v1_attendance_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Month (1-12)",
              "title": "Month"
            },
            "description": "Month (1-12)"
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Year (e.g. 2026)",
              "title": "Year"
            },
            "description": "Year (e.g. 2026)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceMeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "List check-in / attendance records with date filter",
        "description": "Super Admin & HR Admin endpoint to view check-in / attendance history of all employees.\nSupports filtering by exact date, date range (start_date to end_date), employee, department, and status.",
        "operationId": "get_attendance_history_api_v1_attendance_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exact date (YYYY-MM-DD)",
              "title": "Date"
            },
            "description": "Filter by exact date (YYYY-MM-DD)"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter start date (YYYY-MM-DD) inclusive",
              "title": "Start Date"
            },
            "description": "Filter start date (YYYY-MM-DD) inclusive"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter end date (YYYY-MM-DD) inclusive",
              "title": "End Date"
            },
            "description": "Filter end date (YYYY-MM-DD) inclusive"
          },
          {
            "name": "employee_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by employee ID",
              "title": "Employee Id"
            },
            "description": "Filter by employee ID"
          },
          {
            "name": "department_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by department ID",
              "title": "Department Id"
            },
            "description": "Filter by department ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by attendance status",
              "title": "Status"
            },
            "description": "Filter by attendance status"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/history": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "Super Admin view check-in history of all with date filter",
        "description": "Super Admin & HR Admin endpoint to view check-in / attendance history of all employees.\nSupports filtering by exact date, date range (start_date to end_date), employee, department, and status.",
        "operationId": "get_attendance_history_api_v1_attendance_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exact date (YYYY-MM-DD)",
              "title": "Date"
            },
            "description": "Filter by exact date (YYYY-MM-DD)"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter start date (YYYY-MM-DD) inclusive",
              "title": "Start Date"
            },
            "description": "Filter start date (YYYY-MM-DD) inclusive"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter end date (YYYY-MM-DD) inclusive",
              "title": "End Date"
            },
            "description": "Filter end date (YYYY-MM-DD) inclusive"
          },
          {
            "name": "employee_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by employee ID",
              "title": "Employee Id"
            },
            "description": "Filter by employee ID"
          },
          {
            "name": "department_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by department ID",
              "title": "Department Id"
            },
            "description": "Filter by department ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by attendance status",
              "title": "Status"
            },
            "description": "Filter by attendance status"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/regularize": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Submit attendance regularization request",
        "description": "Request attendance regularization for missed or incorrect check-in/out.",
        "operationId": "regularize_attendance_api_v1_attendance_regularize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckInRequest": {
        "properties": {
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "method": {
            "$ref": "#/components/schemas/CheckMethod",
            "default": "mobile_geofence"
          }
        },
        "type": "object",
        "title": "CheckInRequest"
      },
      "CheckInResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "employee_id": {
            "type": "integer",
            "title": "Employee Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "check_in_time": {
            "type": "string",
            "format": "date-time",
            "title": "Check In Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "employee_id",
          "date",
          "check_in_time",
          "status"
        ],
        "title": "CheckInResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}